perf: ⚡️ eliminate unused docker layer, repo and license
intermediate layer scheduled cronjob on ininstalled executable. if we
want to run a cronjob we should do us in a sidecar docker container
rather than using an abstraction layer. this eliminates dependence on
the repo DockerizedMariaDB, the need to include the need to include its
CC0 license in PKC open source page and simplifies our architecture by
removing an abstraction layer.